Refactor command handling
Commands are now turned into events when parsed,
and will be fired on the next iteration of the event loop.
Functions processing commands now also receive
a pointer to the event that called the function.
The event struct contains context about the event,
such as the client that sent a particular command.

The store that stores all of the DB data
no longer uses a mutex, as it is unnecessary when
the program is a single threaded event loop
and each access is contained inside of a single event.

The server now immediately boots into the event loop
after settings are configured and the server socket
is created and bound to the set port.